requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
playwright==1.40.0
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
Pillow==10.0.1
cohere==4.37 